tests/test_wheel_importer.py
import pytest
from fastapi_project.app.services.wheel_service import WheelService
from fastapi_project.app.database import SessionLocal

@pytest.fixture
def db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

def test_import_wheel_cycle_from_csv(db):
    # NOTE: WheelService.import_from_csv does not exist yet.
    # This test is a placeholder and will be replaced once the method is implemented.
    # Tracked in tests.json as WHL-002 / DET-* suites.
    pytest.skip("WheelService.import_from_csv not implemented — tracked in tests.json")
