requirements.txt
# -----------------------------------------------------------------------
# Core runtime dependencies -- pinned to minor version for reproducibility.
# Run `pip install -r requirements.txt` to install.
# Run `pip install -r requirements-dev.txt` to also install dev tools.
# -----------------------------------------------------------------------

# Web framework
fastapi>=0.111.0,<0.116.0
uvicorn[standard]>=0.29.0,<0.35.0

# Database
sqlalchemy>=2.0.30,<2.1.0
alembic>=1.13.0,<1.15.0

# Data validation
pydantic>=2.7.0,<2.10.0
pydantic[email]>=2.7.0,<2.10.0
email-validator>=2.1.1,<2.3.0

# Auth
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
cryptography>=42.0.0,<44.0.0

# HTTP / multipart
requests>=2.31.0,<2.33.0
httpx>=0.27.0,<0.29.0
python-multipart>=0.0.9,<0.0.20

# Market data
yfinance>=0.2.38,<0.2.55
twelvedata>=1.2.11,<1.4.0

# Rate limiting
slowapi>=0.1.9,<0.2.0

# Utilities
python-dotenv>=1.0.1,<1.1.0

# PostgreSQL driver (needed on Render prod; safe to install locally)
psycopg2-binary>=2.9.9,<2.10.0
